requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
Pillow==9.2.0
pipreqs==0.4.11
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
yarg==0.1.9
